Create new directory data/constant, Move fileAppend test to logDaoHelperTest
