+++ src/main/java/com/google/error_analyzer/backend/MockErrorQuery.java
@@ -11,8 +11,8 @@ limitations under the License.*/
 
 package com.google.error_analyzer.backend;
 
-import com.google.error_analyzer.data.Keywords; 
-import com.google.error_analyzer.data.RegexStrings;
+import com.google.error_analyzer.data.constant.Keywords; 
+import com.google.error_analyzer.data.constant.RegexStrings;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -20,7 +20,6 @@ import java.util.regex.Pattern;
 /** 
 * This class mocks the BooleanQuery of elasticsearch.
 */
-
 public class MockErrorQuery {
     private final String keyWordsQueryString = Keywords.getQueryString();
     private final String[] keyWordsList = keyWordsQueryString.split(" OR ");
