+++ src/main/java/com/google/error_analyzer/data/constant/Keywords.java
@@ -9,7 +9,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.google.error_analyzer.data;
+package com.google.error_analyzer.data.constant;
 
 import java.util.ArrayList;
 
@@ -17,9 +17,8 @@ import java.util.ArrayList;
 * This class contains all the keywords used in fulltext query.
 * A log line will apear in search hits if one or more of these terms appear in it.
 */
-
 public class Keywords {
-
+    private final static String LogicString = " OR ";
     private static final ArrayList < String > keywordsList = new ArrayList < String > () {
         {
             add("error");
@@ -30,14 +29,14 @@ public class Keywords {
         }
     };
 
-    //returns kewords as a single string combined together in OR logic.
+    //returns keywords as a single string combined together in OR logic.
     public static String getQueryString() {
         String queryString = "";
         for (String keyword : keywordsList) {
-            if (queryString.length() == 0) {
+            if (queryString.isEmpty()) {
                 queryString = queryString.concat(keyword);
             } else {
-                queryString = queryString.concat(" OR " + keyword);
+                queryString = queryString.concat(LogicString.concat(keyword));
             }
         }
         return queryString;
