+++ src/main/java/com/google/error_analyzer/data/constant/LogFields.java
@@ -9,7 +9,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.google.error_analyzer.data;
+package com.google.error_analyzer.data.constant;
 
 /** Field names in an index file.
 */
